Add files via upload
This is the Registry key that needs to be created to disable Triple DES 168.  The info in KB245030 appears to be incorrect.  If you run a vulnerability scan with the information in the KB the machine will still show vulnerable.  Setting the registry key as it is in this .reg file will pass the vulnerability scan.

As always - test this in a test environment before using it in production.  You should also make sure you have a good backup of your machine and registry.